fix(persistence): move repository errors into folders
