+++ catalog/persistence/src/repositories/errors/category-entity-not-found.error.ts
