+++ catalog/persistence/src/repositories/errors/category-group-entity-not-found.error.ts
