+++ collaboration/persistence/src/repositories/errors/chat-entity-not-found.error.ts
