+++ collaboration/persistence/src/repositories/errors/discussion-entity-not-found.error.ts
