+++ collaboration/persistence/src/repositories/errors/message-entity-not-found.error.ts
