+++ collaboration/persistence/src/repositories/errors/project-entity-not-found.error.ts
