+++ collaboration/persistence/src/repositories/errors/reply-entity-not-found.error.ts
