+++ files/persistence/src/repositories/errors/file-entity-not-found.error.ts
