+++ files/persistence/src/repositories/errors/upload-entity-not-found.error.ts
