+++ portfolio/persistence/src/repositories/errors/portfolio-entity-not-found.error.ts
